feat: add job filter
